backend/requirements.txt
Flask==2.0.1
Werkzeug==2.2.2
flask_mongoengine
flask-cors
Flask-Redis
python-dateutil
